+++ test/native/integration-test-helpers/constants.js
@@ -1,0 +1,1 @@
+export const FRAME_TIME = 1000 / 60;
