+++ test/native/integration-test-helpers/get-editor-html.js
@@ -1,0 +1,37 @@
+/**
+ * WordPress dependencies
+ */
+import {
+	subscribeParentGetHtml,
+	provideToNative_Html as provideToNativeHtml,
+} from '@wordpress/react-native-bridge';
+
+// Set up the mocks for getting the HTML output of the editor
+let triggerHtmlSerialization;
+let serializedHtml;
+subscribeParentGetHtml.mockImplementation( ( callback ) => {
+	if ( ! triggerHtmlSerialization ) {
+		triggerHtmlSerialization = callback;
+		return {
+			remove: () => {
+				triggerHtmlSerialization = undefined;
+			},
+		};
+	}
+} );
+provideToNativeHtml.mockImplementation( ( html ) => {
+	serializedHtml = html;
+} );
+
+/**
+ * Gets the current HTML output of the editor.
+ *
+ * @return {string} HTML output
+ */
+export function getEditorHtml() {
+	if ( ! triggerHtmlSerialization ) {
+		throw new Error( 'HTML serialization trigger is not defined.' );
+	}
+	triggerHtmlSerialization();
+	return serializedHtml;
+}
